requirements.txt
beautifulsoup4==4.11.1
requests==2.28.1
tqdm==4.64.1
pillow==9.3.0
